a sample for newout1.
This a sample extracted from fresh installed ubuntu, the permissions on home directory is messed up. actutally its better to change permissions on home directory based on its name. i cant find other ways. (the case where there are multi users, that would be difficult)(also we can serch for directories named under users in passwd and change permissions according to it. this is just a bare prog and we can improve it for more reliability)